backend/requirements.txt
fastapi>=0.103.1
uvicorn>=0.23.2
python-dotenv>=1.0.0
openai>=1.2.0
anthropic>=0.5.0
pillow>=10.0.0
httpx>=0.24.1
pydantic>=2.3.0
python-multipart>=0.0.6 